// src/lib/eufy-security/p2p/types.ts
export const RequestMessageType = {
    STUN:               Buffer.from([0xF1, 0x00]),
    LOOKUP:             Buffer.from([0xF1, 0x20]),
    LOOKUP_WITH_KEY:    Buffer.from([0xF1, 0x26]),
    LOCAL_LOOKUP:       Buffer.from([0xF1, 0x30]),
    PING:               Buffer.from([0xF1, 0xE0]),
    PONG:               Buffer.from([0xF1, 0xE1]),
    CHECK_CAM:          Buffer.from([0xF1, 0x41]),
    DATA:               Buffer.from([0xF1, 0xD0]),
    ACK:                Buffer.from([0xF1, 0xD1]),
    END:                Buffer.from([0xF1, 0xF0])
}

export const ResponseMessageType = {
    STUN:               Buffer.from([0xF1, 0x01]),
    LOOKUP_RESP:        Buffer.from([0xF1, 0x21]),
    LOOKUP_ADDR:        Buffer.from([0xF1, 0x40]),
    LOCAL_LOOKUP_RESP:  Buffer.from([0xF1, 0x41]),
    END:                Buffer.from([0xF1, 0xF0]),
    PONG:               Buffer.from([0xF1, 0xE1]),
    PING:               Buffer.from([0xF1, 0xE0]),
    CAM_ID:             Buffer.from([0xF1, 0x42]),
    ACK:                Buffer.from([0xF1, 0xD1]),
    DATA:               Buffer.from([0xF1, 0xD0])
}

export const EufyP2PDataType = {
    DATA:               Buffer.from([0xD1, 0x00]),
    VIDEO:              Buffer.from([0xD1, 0x01]),
    CONTROL:            Buffer.from([0xD1, 0x02])
}